+++ crates/astria-composer/src/collectors/mod.rs
@@ -1,5 +1,4 @@
 pub(crate) mod geth;
-mod geth_collector_builder;
 pub(crate) mod grpc;
 
 const EXECUTOR_SEND_TIMEOUT: Duration = Duration::from_millis(500);
@@ -7,5 +6,4 @@ const EXECUTOR_SEND_TIMEOUT: Duration = Duration::from_millis(500);
 use std::time::Duration;
 
 pub(crate) use geth::Geth;
-pub(crate) use geth_collector_builder::GethCollectorBuilder;
 pub(crate) use grpc::Grpc;
