+++ crates/astria-composer/src/lib.rs
@@ -45,7 +45,6 @@ mod composer;
 pub mod config;
 mod executor;
 mod grpc;
-mod grpc_server_builder;
 mod rollup;
 
 pub use build_info::BUILD_INFO;
